+++ save-agent/save-agent-common/src/commonMain/kotlin/com/saveourtool/save/utils/PlatformUtils.kt
@@ -5,6 +5,7 @@
 package com.saveourtool.save.utils
 
 import com.saveourtool.save.core.logging.logDebug
+import com.saveourtool.save.core.utils.GenericAtomicReference
 
 import kotlin.time.Duration
 import kotlin.time.DurationUnit
